UHF-11186: Removed obsolete try catches and combined the CKEditor plugin tests and the test base.
